Ignore .vs build files, fix some warnings from MSVC

+++ include/tl_count.hxx
@@ -6,7 +6,7 @@ namespace pi::tl::internal
     template<typename SearchedType, typename Head, typename ...Tail>
     auto constexpr count()
     {
-        return (std::is_same_v<SearchedType, Head> + ... + std::is_same_v<SearchedType, Tail>);
+        return (static_cast<size_t>(std::is_same_v<SearchedType, Head>) + ... + static_cast<size_t>(std::is_same_v<SearchedType, Tail>));
     }
 }
